Add dump and restore filters
This patch adds an optional argument to dump and restore functions that
can be used for passing options. Currently, there's the only option
available, 'filter', which is supposed to be a function that takes a
space and a tuple and returns a tuple. This function is called for each
dumped/restored tuple. It can be used to overwrite what is written to
the dump file or restored. If it returns nil the tuple will be skipped.

The 'filter' option can be used to convert memtx spaces to vinyl as
shown below:

    dump.restore('dump', {
        filter = function(space, tuple)
            if space.id == box.schema.SPACE_ID then
                return tuple:update{{'=', 4, 'vinyl'}}
            else
                return tuple
            end
        end
    })

Closes #1